Update and rename app_soap.js to soap.js
